% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldBankData.R
\name{WorldBankData_function}
\alias{WorldBankData_function}
\title{A Global Threatened Birds and development Mapping Function}
\usage{
WorldBankData_function(ans = TRUE)
}
\arguments{
\item{Yes}{Do you want to map gdp and birds? Defaults to TRUE.}
}
\description{
This function allows you to visualize global threatened birds and development
}
\examples{
WorldBankData_function()
}
\keyword{gdp}
